Update rock paper scissors.py
First import random module, create a list of choices and take a input from user as well as the computer. Use a while loop for retrieving valid choice from the user and the by using conditional statements proceed with further action. All the program should be written in while loop and break statement in the end, to ask the user to re run the program.